Refactor import statements for consistency and clarity; add Parameters class for dynamics and output functions
