+++ examples/nonlinear_plant_step_response.py
@@ -6,8 +6,9 @@ import os
 import torch
 import numpy as np
 import matplotlib.pyplot as plt
-from torchcontrol.plants import NonlinearSystem
-from torchcontrol.plants import NonlinearSystemCfg, Parameters
+
+from torchcontrol.system import Parameters
+from torchcontrol.plants import NonlinearSystem, NonlinearSystemCfg
 
 def nonlinear_oscillator(x, u, t, params):
     # x: [num_envs, 2], u: [num_envs, 1], t: scalar or [num_envs], params: Parameters
