+++ examples/uav_thrust_descent_to_hover.py
@@ -10,8 +10,8 @@ import torch
 import numpy as np
 import matplotlib.pyplot as plt
 
-from torchcontrol.plants.nonlinear_system import NonlinearSystem
-from torchcontrol.plants.nonlinear_system_cfg import NonlinearSystemCfg, Parameters
+from torchcontrol.system import Parameters
+from torchcontrol.plants import NonlinearSystem, NonlinearSystemCfg
 from torchcontrol.utils.math import quaternion_to_dcm, omega_quat_matrix
 
 def uav_dynamics(x, u, t, params):
