+++ torchcontrol/system/__init__.py
@@ -1,7 +1,9 @@
+from .parameters import Parameters
 from .system_base import SystemBase
 from .system_cfg import SystemCfg, configclass
 
 __all__ = [
+    "Parameters",
     "SystemBase",
     "SystemCfg",
     "configclass"
